requirements.txt
python-vlc
keyboard